move format tests to model tests
