+++ packages/nimble-components/src/table-column/number-text/tests/table-column-number-text.spec.ts
@@ -274,280 +274,4 @@ describe('TableColumnNumberText', () => {
             expect(pageObject.getGroupHeaderTitle(0)).toBe('');
         });
     });
-
-    describe('with default formatting', () => {
-        const testCases: readonly NumericTestCase[] = [
-            {
-                name: 'NEGATIVE_INFINITY renders as -∞',
-                value: Number.NEGATIVE_INFINITY,
-                expectedRenderedString: '-∞'
-            },
-            {
-                name: 'POSITIVE_INFINITY renders as ∞',
-                value: Number.POSITIVE_INFINITY,
-                expectedRenderedString: '∞'
-            },
-            {
-                name: 'NaN renders as NaN',
-                value: Number.NaN,
-                expectedRenderedString: 'NaN'
-            },
-            {
-                name: '-0 renders without negative sign',
-                value: -0,
-                expectedRenderedString: '0'
-            },
-            {
-                name: '+0 renders without positive sign',
-                value: 0,
-                expectedRenderedString: '0'
-            },
-            {
-                name: 'without exponential notation limits to 6 digits with rounding decimals up',
-                value: 1.23456789,
-                expectedRenderedString: '1.23457'
-            },
-            {
-                name: 'without exponential notation limits to 6 digits with rounding decimals down',
-                value: 10.001122,
-                expectedRenderedString: '10.0011'
-            },
-            {
-                name: 'does not use exponential notation for -999,999',
-                value: -999999,
-                expectedRenderedString: '-999,999'
-            },
-            {
-                name: 'does not use exponential notation for -999,999.4999',
-                value: -999999.4999,
-                expectedRenderedString: '-999,999'
-            },
-            {
-                name: 'uses exponential notation for -999,999.5',
-                value: -999999.5,
-                expectedRenderedString: '-1E6'
-            },
-            {
-                name: 'uses exponential notation for -1,000,000',
-                value: -1000000,
-                expectedRenderedString: '-1E6'
-            },
-            {
-                name: 'does not use exponential notation for 999,999',
-                value: 999999,
-                expectedRenderedString: '999,999'
-            },
-            {
-                name: 'does not use exponential notation for 999,999.4999',
-                value: 999999.4999,
-                expectedRenderedString: '999,999'
-            },
-            {
-                name: 'uses exponential notation for 999,999.5',
-                value: 999999.5,
-                expectedRenderedString: '1E6'
-            },
-            {
-                name: 'uses exponential notation for 1,000,000',
-                value: 1000000,
-                expectedRenderedString: '1E6'
-            },
-            {
-                name: 'does not use exponential notation for 0.001',
-                value: 0.001,
-                expectedRenderedString: '0.001'
-            },
-            {
-                name: 'does not use exponential notation for 0.000995',
-                value: 0.000995,
-                expectedRenderedString: '0.001'
-            },
-            {
-                name: 'uses exponential notation for 0.000994',
-                value: 0.000994,
-                expectedRenderedString: '9.94E-4'
-            },
-            {
-                name: 'does not show more than 6 digits even if the rendered value has fewer than 6 significant digits',
-                value: 0.0123456,
-                expectedRenderedString: '0.01235'
-            },
-            {
-                name: 'does not show decimals for an integer value',
-                value: 16,
-                expectedRenderedString: '16'
-            },
-            {
-                name: 'does not add extra decimals to reach 6 digits',
-                value: -98.75,
-                expectedRenderedString: '-98.75'
-            },
-            {
-                name: 'converts numbers with large magnitudes to exponential notation',
-                value: -123456789.123456789,
-                expectedRenderedString: '-1.23457E8'
-            },
-            {
-                name: 'converts numbers with small magnitudes to exponential notation',
-                value: 0.000000123456789,
-                expectedRenderedString: '1.23457E-7'
-            },
-            {
-                name: 'MAX_SAFE_INTEGER + 9999 renders as an exponential with most significant digits of MAX_SAFE_INTEGER',
-                value: Number.MAX_SAFE_INTEGER + 9999,
-                expectedRenderedString: '9.0072E15'
-            },
-            {
-                name: 'MIN_SAFE_INTEGER - 9999 renders as an exponential with most significant digits of MIN_SAFE_INTEGER',
-                value: Number.MIN_SAFE_INTEGER - 9999,
-                expectedRenderedString: '-9.0072E15'
-            }
-        ] as const;
-
-        const focused: string[] = [];
-        const disabled: string[] = [];
-        for (const testCase of testCases) {
-            const specType = getSpecTypeByNamedList(
-                testCase,
-                focused,
-                disabled
-            );
-            // eslint-disable-next-line @typescript-eslint/no-loop-func
-            specType(
-                `${testCase.name}`,
-                // eslint-disable-next-line @typescript-eslint/no-loop-func
-                async () => {
-                    await connect();
-
-                    await element.setData([{ number1: testCase.value }]);
-                    await waitForUpdatesAsync();
-
-                    expect(pageObject.getRenderedCellContent(0, 0)).toBe(
-                        testCase.expectedRenderedString
-                    );
-                    expect(
-                        pageObject.getRenderedGroupHeaderContent(0)
-                    ).toContain(testCase.expectedRenderedString);
-                }
-            );
-        }
-    });
-
-    describe('with round-to-integer formatting', () => {
-        beforeEach(async () => {
-            columnInstances.column1.format = NumberTextFormat.roundToInteger;
-            await waitForUpdatesAsync();
-        });
-
-        const testCases: readonly NumericTestCase[] = [
-            {
-                name: 'NEGATIVE_INFINITY renders as -∞',
-                value: Number.NEGATIVE_INFINITY,
-                expectedRenderedString: '-∞'
-            },
-            {
-                name: 'POSITIVE_INFINITY renders as ∞',
-                value: Number.POSITIVE_INFINITY,
-                expectedRenderedString: '∞'
-            },
-            {
-                name: 'NaN renders as NaN',
-                value: Number.NaN,
-                expectedRenderedString: 'NaN'
-            },
-            {
-                name: '-0 renders without negative sign',
-                value: -0,
-                expectedRenderedString: '0'
-            },
-            {
-                name: '+0 renders without positive sign',
-                value: 0,
-                expectedRenderedString: '0'
-            },
-            {
-                name: '-0.1 renders as 0 without negative sign',
-                value: -0.1,
-                expectedRenderedString: '0'
-            },
-            {
-                name: 'rounds down positive numbers',
-                value: 1.23,
-                expectedRenderedString: '1'
-            },
-            {
-                name: 'rounds up positive numbers',
-                value: 1.76,
-                expectedRenderedString: '2'
-            },
-            {
-                name: 'rounds down negative numbers',
-                value: -1.76,
-                expectedRenderedString: '-2'
-            },
-            {
-                name: 'rounds up negative numbers',
-                value: -1.23,
-                expectedRenderedString: '-1'
-            },
-            {
-                name: 'shows more than 6 digits for positive numbers',
-                value: 987654321,
-                expectedRenderedString: '987,654,321'
-            },
-            {
-                name: 'shows more than 6 digits for negative numbers',
-                value: -123456789,
-                expectedRenderedString: '-123,456,789'
-            },
-            {
-                name: 'MAX_SAFE_INTEGER + 1 renders as the exact value',
-                value: Number.MAX_SAFE_INTEGER + 1,
-                expectedRenderedString: '9,007,199,254,740,992'
-            },
-            {
-                name: 'MAX_SAFE_INTEGER + 1.5  renders with rounding precision errors',
-                value: Number.MAX_SAFE_INTEGER + 1.5,
-                expectedRenderedString: '9,007,199,254,740,992'
-            },
-            {
-                name: 'MIN_SAFE_INTEGER - 1 renders as the exact value',
-                value: Number.MIN_SAFE_INTEGER - 1,
-                expectedRenderedString: '-9,007,199,254,740,992'
-            },
-            {
-                name: 'MIN_SAFE_INTEGER - 1.5  renders with rounding precision errors',
-                value: Number.MIN_SAFE_INTEGER - 1.5,
-                expectedRenderedString: '-9,007,199,254,740,992'
-            }
-        ] as const;
-
-        const focused: string[] = [];
-        const disabled: string[] = [];
-        for (const testCase of testCases) {
-            const specType = getSpecTypeByNamedList(
-                testCase,
-                focused,
-                disabled
-            );
-            // eslint-disable-next-line @typescript-eslint/no-loop-func
-            specType(
-                `${testCase.name}`,
-                // eslint-disable-next-line @typescript-eslint/no-loop-func
-                async () => {
-                    await connect();
-
-                    await element.setData([{ number1: testCase.value }]);
-                    await waitForUpdatesAsync();
-
-                    expect(pageObject.getRenderedCellContent(0, 0)).toBe(
-                        testCase.expectedRenderedString
-                    );
-                    expect(
-                        pageObject.getRenderedGroupHeaderContent(0)
-                    ).toContain(testCase.expectedRenderedString);
-                }
-            );
-        }
-    });
 });
